undid multi generics but bugfixed async.
Generics is not meant to be a major tool in this language. just one of the tools that can be used to build more major tools. This will be the minimum provided by this version of the compiler. 
The compiler will be rewritten in itself in a future version using these base features to bootstrap.